refactor: added references.rs
