+++ src/lib.rs
@@ -70,6 +70,7 @@ mod bytes;
 #[cfg(feature = "malloc_size_of")]
 mod mallocsizeof;
 mod rawsmallvec;
+mod references;
 #[cfg(feature = "serde")]
 mod serde;
 #[cfg(test)]
@@ -90,10 +91,6 @@ use {
         vec::Vec
     },
     core::{
-        borrow::{
-            Borrow,
-            BorrowMut
-        },
         fmt::Debug,
         hash::{
             Hash,
@@ -2069,17 +2066,6 @@ impl<T, const N: usize> Drop for IntoIter<T, N> {
     }
 }
 
-impl<T, const N: usize> core::ops::Deref for SmallVec<T, N> {
-    type Target = [T];
-
-    #[inline]
-    fn deref(&self) -> &Self::Target { self.as_slice() }
-}
-impl<T, const N: usize> core::ops::DerefMut for SmallVec<T, N> {
-    #[inline]
-    fn deref_mut(&mut self) -> &mut Self::Target { self.as_mut_slice() }
-}
-
 /// This function is used in the [`smallvec`] macro.
 /// It is recommended to use the macro instead of using thís function.
 #[doc(hidden)]
@@ -2813,26 +2799,6 @@ where T: Ord
 
 impl<T: Hash, const N: usize> Hash for SmallVec<T, N> {
     fn hash<H: Hasher>(&self, state: &mut H) { self.as_slice().hash(state) }
-}
-
-impl<T, const N: usize> Borrow<[T]> for SmallVec<T, N> {
-    #[inline]
-    fn borrow(&self) -> &[T] { self.as_slice() }
-}
-
-impl<T, const N: usize> BorrowMut<[T]> for SmallVec<T, N> {
-    #[inline]
-    fn borrow_mut(&mut self) -> &mut [T] { self.as_mut_slice() }
-}
-
-impl<T, const N: usize> AsRef<[T]> for SmallVec<T, N> {
-    #[inline]
-    fn as_ref(&self) -> &[T] { self.as_slice() }
-}
-
-impl<T, const N: usize> AsMut<[T]> for SmallVec<T, N> {
-    #[inline]
-    fn as_mut(&mut self) -> &mut [T] { self.as_mut_slice() }
 }
 
 impl<T: Debug, const N: usize> Debug for SmallVec<T, N> {
